code/requirements.txt
dnspython==2.2.1
netifaces==0.11.0
flask==2.2.2 
